battery level Signed-off-by: kevinachoolhun <kevina.choolhun@gmail.com>

+++ src/plotAFriend/PlotAFriendSaver/Model/Logging/SuggestLogger.java
@@ -18,7 +18,7 @@ public class SuggestLogger {
 		return instance;
 	}
 
-	private int getBattery(Context c) {
+	private double getBattery(Context c) {
 		return SuggestBatteryReader.getInstance().getBatteryLevel(c);
 	}
